# docs/api_spec.yaml
openapi: 3.0.0
info:
  title: Face Recognition API
  version: 1.0.0
paths:
...
  /api/models/upload:
    post:
      summary: Upload a new model version for OTA
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelVersion'
      responses:
        '200':
          description: Model uploaded
  /api/models/download:
    get:
      summary: Download model for OTA update
      parameters:
        - name: device_id
          in: query
          schema:
            type: string
        - name: model_version
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Model data
...
  /api/plans:
    get:
      summary: Get all available subscription plans
      responses:
        '200':
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanResponse'
    post:
      summary: Create a new subscription plan (admin only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanResponse'
      responses:
        '200':
          description: Plan created successfully
  /api/plans/{plan_id}:
    get:
      summary: Get a specific plan by ID
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
...
  /api/payments/create-session:
    post:
      summary: Create a Stripe checkout session for payment
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '200':
          description: Checkout session created
  /api/payments/webhook:
    post:
      summary: Handle Stripe webhooks for payment confirmation
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed successfully
  /api/payments/history:
    get:
      summary: Get payment history for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponse'
...
  /api/ai/assistant:
    post:
      summary: Query the AI assistant for help with face recognition tasks
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIAssistantRequest'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAssistantResponse'
  /api/ai/analyze-image:
    post:
      summary: Analyze an uploaded image using AI vision capabilities
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image_data:
                  type: string
      responses:
        '200':
          description: Image analysis results
...
components:
  schemas:
...
    ModelVersion:
      type: object
      properties:
        version_id:
          type: string
        model_data:
          type: string
          format: binary
        created_at:
          type: string
        description:
          type: string
...
    PlanResponse:
      type: object
      properties:
        plan_id:
          type: string
        name:
          type: string
        price:
          type: number
        features:
          type: array
          items:
            type: string
        limits:
          type: object
...
    PaymentCreate:
      type: object
      properties:
        plan_id:
          type: string
        amount:
          type: number
    PaymentResponse:
      type: object
      properties:
        payment_id:
          type: string
        user_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        stripe_payment_id:
          type: string
        created_at:
          type: string
...
    AIAssistantRequest:
      type: object
      properties:
        query:
          type: string
    AIAssistantResponse:
      type: object
      properties:
        query:
          type: string
        response:
          type: string
        model_used:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT